[update] Moves unrelated functionality
Moves Projectile and Environment structs to the ballistics
example where they belong. If they become relevant again they will
be added again.
