+++ examples/vector_ballistics.rs
@@ -1,7 +1,39 @@
 //! Example of vector ballistics
 
-use libray::{graphics, math::Tuple, Environment, Projectile};
+use libray::{graphics, math::Tuple};
 use std::fs;
+
+/// Projectile struct
+#[derive(Debug, Clone, Copy)]
+pub struct Projectile {
+    /// Tuple representing projectile position
+    pub position: Tuple,
+    /// Tuple representing projectile velocity
+    pub velocity: Tuple,
+}
+
+/// Environment struct
+#[derive(Debug, Clone, Copy)]
+pub struct Environment {
+    /// Tuple representing environment gravity
+    pub gravity: Tuple,
+    /// Tuple representing environment gravity wind
+    pub wind: Tuple,
+}
+
+impl Environment {
+    /// Tick environment one step
+    ///
+    /// # Arguments:
+    ///
+    /// * `projectile` - Projectile to be affected by environment properties
+    pub fn tick(&self, projectile: Projectile) -> Projectile {
+        let position = projectile.position + projectile.velocity;
+        let velocity = projectile.velocity + self.gravity + self.wind;
+
+        Projectile { position, velocity }
+    }
+}
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut p = Projectile {
