+++ src/lib.rs
@@ -18,37 +18,3 @@
 
 pub mod graphics;
 pub mod math;
-
-use math::Tuple;
-
-/// Projectile struct
-#[derive(Debug, Clone, Copy)]
-pub struct Projectile {
-    /// Tuple representing projectile position
-    pub position: Tuple,
-    /// Tuple representing projectile velocity
-    pub velocity: Tuple,
-}
-
-/// Environment struct
-#[derive(Debug, Clone, Copy)]
-pub struct Environment {
-    /// Tuple representing environment gravity
-    pub gravity: Tuple,
-    /// Tuple representing environment gravity wind
-    pub wind: Tuple,
-}
-
-impl Environment {
-    /// Tick environment one step
-    ///
-    /// # Arguments:
-    ///
-    /// * `projectile` - Projectile to be affected by environment properties
-    pub fn tick(&self, projectile: Projectile) -> Projectile {
-        let position = projectile.position + projectile.velocity;
-        let velocity = projectile.velocity + self.gravity + self.wind;
-
-        Projectile { position, velocity }
-    }
-}
